Refactor policy css
