+++ site/components/policy/policy-box/style.css
@@ -1,0 +1,38 @@
+@value mediumScreen from "../../../css/_sizes.css";
+
+.box {
+  display: flex;
+  align-items: baseline;
+}
+
+.numberContainer {
+  margin-right: 30px;
+}
+
+.number {
+  composes: boldSansSerif from '../../../css/typography/_fonts.css';
+  font-size: 16px;
+  line-height: 1.1;
+  padding: 0;
+  border-bottom: 5px solid #FC1D42;
+  display: block;
+  position: relative;
+  top: -2px;
+}
+
+.policyText {
+  max-width: 610px;
+  width: 100%;
+  margin-bottom: 10px;
+}
+
+.policyText p b {
+  font-family: 'ProximaNova-Bold', sans-serif;
+}
+
+@media mediumScreen {
+  .number {
+    top: -9px;
+  }
+}
+
