+++ site/pages/privacy-policy/index.js
@@ -1,12 +1,11 @@
 // @flow
-/* eslint-disable react/no-unescaped-entities */
 
 import React, { Fragment } from 'react';
 import Social from '../../components/social';
-import { H1, H2 } from '../../components/text';
+import { H1 } from '../../components/text';
+import PolicyBox from '../../components/policy/policy-box';
 import metaImage from '../home/meta-image.jpg';
 import styles from './style.css';
-
 import policies from './policy';
 
 const social = {
@@ -16,72 +15,6 @@ const social = {
   metaImage,
   url: 'cookie-policy',
 };
-
-type StatementT = {
-  body: Function,
-};
-
-type StatementWithIndexes = {
-  statementIndex: number,
-  policyIndex: number,
-  body: Function,
-};
-
-type Policy = {
-  policyIndex: number,
-  heading: string,
-  body?: Function,
-  statements: Array<StatementT>,
-};
-
-const Statement = ({ body, statementIndex, policyIndex }: StatementWithIndexes) => (
-  <li className={styles.box}>
-    <div className={styles.numberContainer}>
-      <span className={styles.statementNumber}>{`${policyIndex}.${statementIndex}`}</span>
-    </div>
-    <div className={styles.policyText}>{body()}</div>
-  </li>
-);
-
-const padNumber = (n: number) => (n < 10 ? '0' + n : +n);
-
-const Statements = ({
-  statements,
-  policyIndex,
-}: {
-  statements: Array<StatementT>,
-  policyIndex: number,
-}) => {
-  return (
-    <ol>
-      {statements.map((statement, statementIndex) => (
-        <Statement
-          key={statementIndex}
-          statementIndex={statementIndex + 1}
-          policyIndex={policyIndex}
-          {...statement}
-        />
-      ))}
-    </ol>
-  );
-};
-
-const PolicyBox = ({ heading, body, statements, policyIndex }: Policy) => (
-  <li>
-    <div className={styles.box}>
-      <div className={styles.numberContainer}>
-        <span className={styles.number}>{padNumber(policyIndex)}</span>
-      </div>
-      <div>
-        <H2 type="fontM2" customClass={styles.mb10}>
-          {heading}
-        </H2>
-        {body && body()}
-      </div>
-    </div>
-    {statements.length > 0 && <Statements statements={statements} policyIndex={policyIndex} />}
-  </li>
-);
 
 const PrivacyPolicyPage = () => {
   return (
